libopenzwave: specify "str" instead of just char*
char* maps to what Cython calls "bytes", which results in the wrong
thing on Python 3. In order to behave as desired on both Pythons, we
add a couple of little wrapper functions to make sure that "str"
becomes char* and char* becomes "str", no matter what "str" currently
means. We assume that openzwave uses utf-8 internally, since we have
to use an encoding to go from bytes to unicode.