Adds better Visual Studio support.
The distutils module is broken in any Pytho releases that are below 3.4, and becuase of this it is not able to build properly. This has been fixed.

Added automatic downlaoding of openzwave
Added modifying the VS solution for x64 builds
Added upgrading the solution to the version of VS that is used
Added setting up a proper VS build environment

Changed how information is printed to the console. This was done so the user would not think the install got hung up. it gives notification that itis still working.