Fixes typo that should be the cause of the build failures
Removes downloading openzwave from appveyor script. I think I need to add the downloading into pyozw_setup for the linux builds not sure yet.

Removes the copying of the VS2010 directory as this does not need to be done.

Popen returns True/False based on whether or not there was error information output. I am not sure if this will halt the program or not.

Streamlined the code a bit more