new build system for windows.
allow use of the following Windows C++ compilers
Visual Studio 2008 - to current
Visual C++ 2008 - to current
Visual C++ 2008 Build Tools - to current
Windows SDK 6.0A - to current
.NET 2.0 - to current

We are not limited to only building with VS 2017 anymore.

I also removed the need for being dependant on setuptools. Since setuptools
versions cause major changes in what can be detected in the way of
MSVC compilers i felt it best to write a whole (complete) detection
system. one that does not use setuptools at all.

I removed parsing of a solution file. as this was dependant on having
MSBuild installed. and also could lead to a nightmare of issues depending
on if the solution file was changed or the version of Visual Studio changed.

I now use distutils to build openzwave. This has made the building
process a whole lot faster and less complex without the need for
parsing any files.

Newer versions of cython will throw out a warning of the language level
is not set. I  set this level in the libopenzwave.pyx file to eliminate
this warning. All MSVC compiler warning have been suppressed to provide a
nice clean output.

The whole build system for Windows is not limited to Windows. It can be
used for the rest of the OS's read the comments in pyozw_common.py for
instructions on how to do this.