adds better handling of strings/bytes/unicode
currently in the libopenzwave.pyx file there are default data types
for str that are set for some of the function/method parameters.
This causes an issue with the test build when running on python 2.
Because a string literal in python 3 by default is unicode (bytes)
everything works as advertised. in python 2 string literals are of
type str(). and because of the marker at the top of the pyx file

#cython: c_string_type=unicode, c_string_encoding=utf8

this is forcing all of the str data types to be unicode.
Now I could go and write a whole second test routine specifically for
python 2. But to me this seems like it is not the proper way to go
about it. Python 2 does not have the best mechanism for dealing with
unicode strings and it can be very confusing. also because everything
is defaulted to str() in python 2. I thought it would be best to grab
the bull by the horns and handle all of the unicode mess internal to
libopenzwave. This is going to make it much easier for the user to be
able to use the library with out having to deal with a bunch of
additional code. especially if their program supports python2 and 3.

It is a real headache to have to decode a path returned by the os module
for something trivial like a logfile location. and if the user is
supporting python 2 and 3 it would add a very large amount of code to
have to deal with issues like this for each and every single
function/method call to python-openzwave.

So what i did was I removed the str data type from the parameters.
I created a function that will check the data type based on the
python version and if it is not either unicode or bytes it will do the
necessary encoding or decoding to make the data the correct type.

The added benefit of doing this is now all of the tests work
properly for python 2 without having to recode the tests.

The API has not changed other then now you can pass str, bytes or unicode
to the functions\methods that would only allow unicode before.

so no changes are made that will affect a users program at all.

the users will be happy. we are happy (no test rewrite to be done).
seems like a win win to me.

I also took into account setting variables that are of type String as
well.