makes code appveyor compatible.
appveyor uses VM's that were set up through an automated process.
when the developers over at appveyor did this they did not copy the full
registry structure of the build tools. I had to  make adjustments to
allow for locating the needed build tools to compile python-openzwave
on windows.

THIS PR IS GOING TO FAIL THE TESTS

this is because of a unicode/string/bytes issue with libopenzwave.pyx
I have submitted a PR #153 that fixes these issues. That pr will need
to be merged in order for this PR to be able to pass the tests.

I suggest setting up a branch. and merge the changes in #153 and also
the changes in this PR check to make sure everything is good to go. then
create a PR to the master branch with that tst branch and merge that way.