adds expanding of `$`, `%` and `~` path variables
when creating the Options class instance the user has the ability to
specify the location of the user path and config path. I have seen on
numerous occasions (specifically with NIX OS's) the user having a
`~` or `$` in the path. Ths are convenient to use but unless the
variables are expanded th paths ar going to be incorrect. So I added
checking for these variables and performing any needed expansion.

NOTE: `~` in Windows serves the same purpose as in NIX OS's. it is a
pointer to the currently logged in user home directory
(profile directory on Windows)