Adds automatic detection of the serial port the zstick is connected on.
This currently only works for Silicon Labs/Sigma Designs chipsets
I do not know the hardware id's of any other chipset manufacturers.

It uses pyserial to enumerate the available ports on the OS and then
checks the hardware ID of the port. pyserial is not listed as a
requirement as it would be an optional install the program tries to
import pyserial and if it cannot import it then it will continue
along it's way.

the auto detection gets activated if there is no supplied `device` to
ZWaveOption.

Windows like most operating systems do function the same in terms of a
path to the serial device. on NIX based systems it is a literal file
path where as in Windows it is a virtual path (namespace object).
Unless the user is very familiar with th Windows OS they are not going
to know to prefix the "COM" port with "\\\\.\\". I added in the
prefixing of the "COM" port if it has not been done.

These additions will make it easier for the user to get up and running.