created a file structure for all of the command classes added some additional methods. remove a lot of repeat iteration code added __getitem__ to the command classes to obtain value(s) added __iter__ to return value.values()
The __getitem__ works in a multitude of ways.

* node[value.value_id]
we can specify a single value_id and that is what will get returned if
available otherwise IndexError is raised
* node[(value.value_id, COMMAND_CLASS)]
a tuple that contains the value id and the command class.
this is going to search for a value id and the value must belong to
the command class otherwise IndexError is raised
* node[value.label]
I tossed in checking the label of thee value as well KeyError is
raised if not found
* node[(value.label, COMMAND_CLASS)]
tuple with the label and the command class. raises KeyError
* node[1: 3]
slicing. I organize all of the indexes for the values in ascending order
then create a list of values in that same order and return the slice
* node[(slice(1, 3), COMMAND_CLASS
same as above except remove any values that are not in the command class
* node[(None, None)]
this is the same thing that __iter__ does
* node[(None, COMMAND_CLASS)]
this returns all values in a command class or an empty list
* node[-1]
this returns the lowest indexed value raises IndexError
* node[(-1, COMMAND_CLASS)] returns the lowest indexed value in a
command class raises IndexError