Adds module dependencies to the setup program
This is a pretty large thing. There is no longer a need to have the
user install any kind of a module (unless wanting to run the unit tests)

The setup program now handles all deps correctly. It also does not create
a seperate folder in the user site-packages folder for each of the different
python-openzwave modules. They are now all contained into a single .egg folder.
It is not a zip. it is a folder. This makes it easier for the user if
they need to remove python-openzwave manually for some reason. simply delete
the singlke folder and remove the line from easy-install.pth that
pointed to that folder.

Handling of installing the proper event dispatcher is also done internally
It is now all seamless to the user.