adds better handling when writing the config files during install
The original design was to delete any existing config files blindly
when doing an install. This is not a good thing because if a user has
added any additional files or made any changes to any of the files
it would delete their work with no way to recover them.

The new system I designed makes a backup of the config files then
installs openzwave. then it places the backup back into place. You have
2 options. the first one is you can set the command line switch so the
config files do not get written. Or you can let it do it's update. If
you go with option 2, the program check to see if the file exists already.
if it does it opens both files and check to see if they are the same.
if they are the same it moves on. If not then it checks for the
revision of the file. If the one that is being installed has a higher
revision, the old one is overwritten. if they are the same revision the
program the goes with whichever file is larger. if there is no revision
in either of the files it goes with the larger file.